+++ examples/src/json_serialization_traits_examples.cpp
@@ -4,11 +4,11 @@
 #include <iostream>
 #include <map>
 #include <tuple>
-#include <jsoncons/json_stream_traits.hpp>
+#include <jsoncons/json_serialization_traits.hpp>
 
 using namespace jsoncons;
 
-namespace examples { namespace streaming {
+namespace examples { namespace serialization_traits {
 
 class Employee
 {
@@ -53,12 +53,12 @@ public:
 };
 
 }}
-using namespace examples::streaming;
+using namespace examples::serialization_traits;
 
 namespace jsoncons
 {
     template <>
-    struct json_stream_traits<std::shared_ptr<Employee>>
+    struct json_serialization_traits<std::shared_ptr<Employee>>
     {
         static void encode(const std::shared_ptr<Employee>& val, json_output_handler& handler)
         {